add authorization & console command create user

+++ app/Models/Repositories/Product.js
@@ -8,24 +8,39 @@ class Product {
     return Promise.all(attrs);
   }
 
-  static async addProduct({ name, user_id: userId, type_id: typeId, price, attributes }) {
-    const product = await this.create({ name, user_id: userId, type_id: typeId, price });
+  static async addProduct({ name, user, response, type_id: typeId, price, attributes }) {
+    const product = await this.create({ name, user_id: user.id, type_id: typeId, price });
     product.attributes = await Product.attachAttrs(product, attributes);
+    response.status(201);
     return product;
   }
 
-  static async updateProduct({ id, name, user_id: userId, type_id: typeId, price, attributes }) {
+  static async updateProduct({ user, response, id, name, type_id: typeId, price, attributes }) {
     const product = await this.findOrFail(id);
-    await product.merge({ name, user_id: userId, type_id: typeId, price });
+
+    const { slug: role } = await user.roles().first();
+    if (role !== 'admin' && product.user_id !== user.id) {
+      return response.status(403).send("You don't have permission rights");
+    }
+
+    await product.merge({ name, user_id: user.id, type_id: typeId, price });
     await product.save();
 
     await Product.attachAttrs(product, attributes);
+    response.status(201);
     return this.findProduct({ id });
   }
 
-  static async deleteProduct({ id }) {
+  static async deleteProduct({ id, user, response }) {
     const product = await this.findOrFail(id);
+
+    const { slug: role } = await user.roles().first();
+    if (role !== 'admin' && product.user_id !== user.id) {
+      return response.status(403).send("You don't have permission rights");
+    }
+
     await product.delete();
+    return response.status(204).json(null);
   }
 
   static async findAllProducts({
